Change lib to src
