+++ composer.json
@@ -2,7 +2,9 @@
   "name": "mpay24/mpay24-php",
   "description": "Offical PHP SDK for SOAP Bindings",
   "autoload": {
-    "psr-4": { "mPay24\\" : "lib/" }
+    "psr-4": {
+      "Mpay24\\": "src/"
+    }
   },
   "require": {
     "php": ">=5.3.3",
