+++ examples/seamless.php
@@ -1,8 +1,8 @@
 <?php
 require("../bootstrap.php");
-use mPay24\MPAY24;
+use mPay24\Mpay24Soap;
 
-$mpay24 = new MPAY24();
+$mpay24 = new Mpay24Soap();
 
 // Each Line is Optional so only add the lines that you need
 $tokenizerConfig = array(
